Add RGBColorMode to EnableInfo
When used in After Effects, the plug-in was only showing Trillions of
Colors as an output option, not Millions. This is a bug in AE that goes
back several versions at least.

But then DWilk came down from heaven and told me about RGBColorMode,
which brings back Millions in After Effects while not appearing to
enable SuperPNG in any Photoshop mode where it shouldn't.