Switch projects to pngquant-msvc-fnordware
It appears that Kornel deleted his msvc branch and then began a new one,
effectively killing the branch I was using as a submodule. I've
resurrected it in my own fork, added that as a submodule, and pointed
the projects to it.

The pngquant-msvc submodule now points to Kornel's latest version of
that branch, which I have not been able to build in vc12 (Visual Studio
Express 2013).

It would be good to get everything to build with Kornel's latest.